+++ app/crud/permission_type.py
@@ -3,11 +3,19 @@ from app.models import PermissionType
 from app.schemas import PermissionTypeCreate, PermissionTypeUpdate
 from datetime import datetime
 from fastapi import HTTPException, status
-from app.utils import get_internal_server_error, get_permission_type_or_404
+from app.utils import get_internal_server_error, get_permission_type_or_404, get_current_user
+from app.utils.permission_checking import check_permission
 
 
-async def create_permission_type(db: Session, request: PermissionTypeCreate):
+
+async def create_permission_type(db: Session, request: PermissionTypeCreate, current_user: dict):
 	try:
+		permission = await check_permission(db, 'manage', 'user', current_user['role_id'])
+		if not permission:
+			return {
+				'mess' : "You don't have permission for accessing this function !",
+				'status_code' : status.HTTP_403_FORBIDDEN 
+			}
 		check = db.query(PermissionType).filter(PermissionType.permission_type_name == request.permission_type_name).first()
 		if check:
 			return {
@@ -40,8 +48,14 @@ async def get_permission_types(db: Session):
 	except Exception as ex:
 		return get_internal_server_error(ex)
 	
-async def update_permission_type(db: Session, request: PermissionTypeUpdate):
+async def update_permission_type(db: Session, request: PermissionTypeUpdate, current_user: dict):
 	try:
+		permission = await check_permission(db, 'manage', 'user', current_user['role_id'])
+		if not permission:
+			return {
+				'mess' : "You don't have permission for accessing this function !",
+				'status_code' : status.HTTP_403_FORBIDDEN 
+			}
 		check = get_permission_type_or_404(db, request.permission_type_id)
 		check.permission_type_name = request.permission_type_name or check.permission_type_name
 		check.description = request.description or check.description
@@ -57,8 +71,14 @@ async def update_permission_type(db: Session, request: PermissionTypeUpdate):
 	except Exception as ex:
 		return get_internal_server_error(ex)
 
-async def delete_permission_type(db: Session, request_id: int):
+async def delete_permission_type(db: Session, request_id: int, current_user: dict):
 	try:
+		permission = await check_permission(db, 'manage', 'user', current_user['role_id'])
+		if not permission:
+			return {
+				'mess' : "You don't have permission for accessing this function !",
+				'status_code' : status.HTTP_403_FORBIDDEN 
+			}
 		check = get_permission_type_or_404(db, request_id)
 		db.delete(check)
 		db.commit()
